requirements.txt
# Core dependencies for the arbitrage monitoring bot
aiohttp>=3.9.0
aiogram>=3.0.0
ccxt>=4.0.0

# Optional but recommended
python-dotenv>=1.0.0

# Additional dependencies for enhanced functionality
anthropic>=0.25.0  # For Claude AI integration
requests>=2.31.0   # For additional API calls
websockets>=12.0   # For real-time updates
pydantic>=2.5.0    # For data validation
uvicorn>=0.24.0    # For API server
fastapi>=0.104.0   # For REST API endpoints

# Database dependencies for contract resolver
sqlalchemy>=2.0.0  # SQL toolkit and ORM
psycopg2-binary>=2.9.0  # PostgreSQL adapter
asyncpg>=0.29.0  # Async PostgreSQL driver (optional, for async operations)



